+++ src/main/java/net/silentchaos512/lib/collection/StackList.java
@@ -1,25 +1,8 @@
-/*
- * SilentLib - StackList
- * Copyright (C) 2018 SilentChaos512
- *
- * This library is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
 package net.silentchaos512.lib.collection;
 
 import net.minecraft.world.Container;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.CraftingInput;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -59,6 +42,14 @@ public final class StackList extends ArrayList<ItemStack> {
         StackList newList = new StackList();
         for (int i = 0; i < inventory.getContainerSize(); ++i) {
             newList.add(inventory.getItem(i));
+        }
+        return newList;
+    }
+
+    public static StackList from(CraftingInput input) {
+        StackList newList = new StackList();
+        for (int i = 0; i < input.size(); ++i) {
+            newList.add(input.getItem(i));
         }
         return newList;
     }